Make metrics table use BIGINT
Folder size may not fit otherwise!

+++ backend/database.sql
@@ -80,7 +80,7 @@ CREATE TABLE IF NOT EXISTS metrics (
   datasource         text,
   board              text,
   date               text,
-  count              integer
+  count              BIGINT
 );
 
 CREATE UNIQUE INDEX IF NOT EXISTS unique_metrics
